[RemoteApp] Refactor toolkit dumping code
Dumping the toolkit (ScubaDiver, Injector, Bootstrap) now dumps to one specific folder in AppData/Local.
It also does not try to overwrite previous versions of files in this location if they haven't changed.
This allows us to Inject multiple times even when an injected process still have open handles to some of the files there.