Hash-Code-assisted dumping fallback
This helps with cases where the `QueryInstances` and `GetRemoteObject` are too far apart so that a GC collection happens.
Previously we'd spot the object is not at it's expected location and return an error.
Now there's also a fallback logic in place so that we are searching other instances of the requested type in the new heap layout, and given a known hash code of the candidate we can try to find a single object with that hash code.
If "GetHashCode" was overridden and there are several objects returning the expected values neither is returned.