[ScubaDiver.API] Force re-build every time.
This forces ScubaDiver.API to rebuild and also RemoteNET.
We actually only want to rebuild RemoteNET everytime but the "dotnet clean" trick can't be used in it's pre-build events
since it cleans also it's dependencies.
This caused the ScuabDiver.API's bin dir to be emptied AFTER ScubaDiver.API completed compiling and right before RemoteNET started compiling - resulting in a "missing DLL" error.