[HarmonyWrapper] Impl Workaround for ref struct parameter
We can't hook a method with a "ref struct" parameter since the parameters in our hooks are of type `object`, ref structs can't be boxed into objects so the Harmony can't add the hook.
This change just avoid those specific parameters - we use new types of hooks that skip the specific problematic parameter.
For example, if this is the target function:
`void Write(Span<byte> a, int b)`
We'll use this overload:
`void UnifiedHook_0100000000(MethodBase __originalMethod, object __instance, object __1)`
(Skipping the first parameter, usually accessed via `__0`)